services/ryan-labeler/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
anthropic==0.40.0
apscheduler==3.10.4
